+++ src/components/feedback/Modal.test.tsx
@@ -11,7 +11,7 @@ import { screen, waitFor } from '@testing-library/react';
 import { Modal, ModalProps } from './Modal';
 import { setup } from '../../test-utils';
 import { Button } from '../basic/Button';
-import { Text } from '../basic/Text';
+import { Text } from '../basic/text/Text';
 
 const ModalTester = ({ children, ...props }: ModalProps): React.JSX.Element => {
 	const [open, setOpen] = useState(false);