+++ src/components/feedback/snackbar/Snackbar.tsx
@@ -10,8 +10,8 @@ import styled, { css, DefaultTheme, keyframes, SimpleInterpolation } from 'style
 
 import { ScreenMode, useScreenMode } from '../../../hooks/useScreenMode';
 import { Button } from '../../basic/Button';
-import { Icon } from '../../basic/Icon';
-import { Text } from '../../basic/Text';
+import { Icon } from '../../basic/icon/Icon';
+import { Text } from '../../basic/text/Text';
 import { TIMERS } from '../../constants';
 import { Container, ContainerProps } from '../../layout/Container';
 import { Row } from '../../layout/Row';