+++ src/components/inputs/Select.test.tsx
@@ -114,7 +114,6 @@ describe('Select', () => {
 				);
 				await user.click(dropdownItem);
 
-				expect(onChange).toHaveBeenCalled();
 				expect(onChange).toHaveBeenCalledTimes(1);
 				expect(onChange).toHaveBeenCalledWith(items[1].value);
 			});
@@ -206,8 +205,7 @@ describe('Select', () => {
 				);
 				await user.click(dropdownItem);
 
-				expect(onChange).toBeCalled();
-				expect(onChange).toBeCalledTimes(1);
+				expect(onChange).toHaveBeenCalledTimes(1);
 				expect(onChange).toHaveBeenCalledWith(items[1].value);
 			});
 			test('click on an item automatically update the value of the select', async () => {