+++ src/components/inputs/commons/InputContainer.tsx
@@ -11,9 +11,7 @@ import { Container, ContainerProps } from '../../layout/Container';
 export const InputContainer = styled(Container)<{
 	background: NonNullable<ContainerProps['background']>;
 	$disabled?: boolean;
-	$hasLabel?: boolean;
 }>`
-	position: relative;
 	${({ $disabled, background, theme }): SimpleInterpolation =>
 		$disabled
 			? css`
@@ -23,8 +21,6 @@ export const InputContainer = styled(Container)<{
 					cursor: text;
 					${pseudoClasses(theme, background)}
 				`};
-	padding: ${({ $hasLabel }): string => ($hasLabel ? '0.0625rem' : '0.625rem')} 0.75rem;
-	gap: 0.5rem;
 	min-height: calc(
 		${({ theme }): string => theme.sizes.font.medium} * 1.5 +
 			${({ theme }): string => theme.sizes.font.extrasmall} * 1.5 + 0.125rem