+++ src/components/navigation/TabBar.test.tsx
@@ -10,7 +10,7 @@ import { screen, within } from '@testing-library/react';
 
 import { DefaultTabBarItem, DefaultTabBarItemProps, TabBar } from './TabBar';
 import { setup } from '../../test-utils';
-import { Text } from '../basic/Text';
+import { Text } from '../basic/text/Text';
 
 describe('TabBar', () => {
 	test('The visually selected element always reflects the selected TabBar prop', async () => {