+++ src/components/navigation/TabBar.tsx
@@ -13,7 +13,7 @@ import { useCombinedRefs } from '../../hooks/useCombinedRefs';
 import { getKeyboardPreset, useKeyboard } from '../../hooks/useKeyboard';
 import { getColor } from '../../theme/theme-utils';
 import { AnyColor } from '../../types/utils';
-import { Text } from '../basic/Text';
+import { Text } from '../basic/text/Text';
 import { Container, ContainerProps } from '../layout/Container';
 
 const CustomText = styled(Text)`